+++ .github/ISSUE_TEMPLATE/product-feature-from-the-roadmap.md
@@ -1,6 +1,6 @@
 ---
-name: Product feature from the Roadmap
-about: Building Stability App by Roadmap
+name: Product feature from the Development
+about: Building Stability App by Development
 title: ''
 labels: ''
 assignees: ''
