+++ src/components/view/index.js
@@ -1,5 +1,5 @@
 export {default as Home} from './Home'
-export {default as Roadmap} from './Roadmap'
+export {default as Development} from './Development'
 export {default as Pools} from './Pools'
 export {default as Generation} from './Generation'
 export {default as Governance} from './Governance'
