+++ src/pages/development.js
@@ -1,0 +1,7 @@
+import {Development } from '../components/view'
+
+function development() {
+    return <Development/>
+}
+
+export default development
