--- src/pages/roadmap.js
@@ -1,7 +1,0 @@
-import {Roadmap } from '../components/view'
-
-function roadmap() {
-    return <Roadmap/>
-}
-
-export default roadmap
